server/requirements.txt
Flask==0.12.2
#hologram-python==0.7.2
paho-mqtt==1.3.1
python-sdk-auth==0.2.0
